refactor: useImageSearch 커스텀 훅 구현
