+++ hooks/src/components/ImageGallery.tsx
@@ -1,37 +1,9 @@
-import React, { useState, useEffect } from "react";
-
-const API_KEY = import.meta.env.VITE_PIXABAY_API_KEY;
-
-interface Image {
-  id: number;
-  webformatURL: string;
-  tags: string;
-}
+import React, { useState } from "react";
+import useImageSearch from "../hooks/useImageSearch";
 
 const ImageGallery: React.FC = () => {
-  const [images, setImages] = useState<Image[]>([]);
   const [keyword, setKeyword] = useState("");
-
-  // useEffect를 사용하여 keyword가 변경될 때마다 이미지를 가져오는 로직을 작성해보세요.
-  useEffect(() => {
-    // 1. fetchImages 함수를 정의하세요.
-    // 2. fetchImages 함수 내부에서 Pixabay API를 호출하여 이미지를 가져오세요.
-    const fetchImages = async () => {
-      try {
-        const res = await fetch(
-          `https://pixabay.com/api/?key=${API_KEY}&q=${keyword}`
-        );
-        const result = await res.json();
-        // 3. 가져온 이미지를 상태(images)에 업데이트하세요.
-        setImages(result.hits);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    // 4. keyword가 존재할 때만 fetchImages 함수를 호출하세요.
-    if (keyword) fetchImages();
-  }, [keyword]);
+  const { images } = useImageSearch(keyword);
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
